+++ src/client/common/types.ts
@@ -4,6 +4,7 @@
 
 import { HexBase64Latin1Encoding } from 'crypto';
 import { Socket } from 'net';
+import { Request as RequestResult } from 'request';
 import { ConfigurationTarget, DiagnosticSeverity, Disposable, DocumentSymbolProvider, Event, Extension, ExtensionContext, OutputChannel, Uri, WorkspaceEdit } from 'vscode';
 import { CommandsWithoutArgs } from './application/commands';
 import { EnvironmentVariables } from './variables/types';
@@ -332,6 +333,17 @@ export const ISocketServer = Symbol('ISocketServer');
 export interface ISocketServer extends Disposable {
     readonly client: Promise<Socket>;
     Start(options?: { port?: number; host?: string }): Promise<number>;
+}
+
+export const IHttpClient = Symbol('IHttpClient');
+export interface IHttpClient {
+    downloadFile(uri: string): Promise<RequestResult>;
+    /**
+     * Downloads file from uri as string and parses them into JSON objects
+     * @param uri The uri to download the JSON from
+     * @param strict Set `false` to allow trailing comma and comments in the JSON, defaults to `true`
+     */
+    getJSON<T>(uri: string, strict?: boolean): Promise<T>;
 }
 
 export const IExtensionContext = Symbol('ExtensionContext');
